+++ tests/features/steps/steps.py
@@ -1,1 +1,0 @@
-from sdklib.test.behave.steps import *