+++ sdklib/http/session.py
@@ -17,7 +17,7 @@ class Cookie(object):
     def load_from_headers(self, headers):
         if not headers:
             return
-        set_cookie_header = headers.get("Set-Cookie")
+        set_cookie_header = headers.get("Set-Cookie", None)
         if set_cookie_header:
             ck = cookies.SimpleCookie()
             ck.load(set_cookie_header)
@@ -26,7 +26,7 @@ class Cookie(object):
     def as_cookie_header_value(self):
         if self.is_empty():
             return ""
-        items = list(self.items())
+        items = list(self._cookie.items())
         name, morsel = items[0]
         output = "%s=%s" % (name, morsel.value)
         for name, morsel in items[1:]:
@@ -35,7 +35,7 @@ class Cookie(object):
         return output
 
     def is_empty(self):
-        return (self._cookie is None) or (self._cookie.items() == [])
+        return (self._cookie is None) or (len(self._cookie.items()) == 0)
 
     def getcookie(self):
         return self._cookie